generative-ai/learning_llm_genai/requirement.txt
# Add the openai dependency for python
openai
python-dotenv
pytest
langchain-openai